fastapi_backend/pyproject.toml
[project]
name = "ai-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "firecrawl-py>=4.8.0",
    "flask>=3.1.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters>=1.0.0",
    "markdownify>=1.2.2",
    "openai>=2.7.2",
    "pandas>=2.3.3",
    "playwright>=1.56.0",
    "pypdf>=6.2.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
    "tiktoken>=0.12.0",
    "twilio>=9.8.6",
    "unstructured[all-docs]>=0.18.20",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
